Rename AuthenicatedRoute to SetupRequiredRoute
