+++ frontend/app/mixins/setup-required-route.js
