+++ frontend/app/routes/application.js
@@ -1,4 +1,4 @@
 import Ember from 'ember';
-import AuthenicatedRoute from '../mixins/authenticated-route';
+import AuthenicatedRoute from '../mixins/setup-required-route';
 
 export default Ember.Route.extend(AuthenicatedRoute);
